Move learning unit services into learning unit app
