+++ learning_unit/services/learning_unit.py
@@ -31,6 +31,7 @@ import osis_learning_unit_sdk
 import urllib3
 from django.conf import settings
 from osis_learning_unit_sdk.api import learning_units_api
+from osis_learning_unit_sdk.model.effective_class import EffectiveClass
 
 from base.models.person import Person
 from frontoffice.settings.osis_sdk import learning_unit as learning_unit_sdk, utils
@@ -49,7 +50,7 @@ class LearningUnitService:
                 learning_units = api_instance.learningunits_list(
                     learning_unit_codes=','.join({str(elem) for elem in learning_unit_codes}),
                     year=year,
-                    ** utils.build_mandatory_auth_headers(person),
+                    **utils.build_mandatory_auth_headers(person),
                 )
                 return learning_units.get('results', {})
             except (osis_learning_unit_sdk.ApiException, urllib3.exceptions.HTTPError,) as e:
@@ -58,3 +59,37 @@ class LearningUnitService:
                 attrs = {'result': None, 'error': e}
                 learning_units = SimpleNamespace(**attrs, attribute_map=attrs)
         return learning_units
+
+    @staticmethod
+    def get_learning_unit_title(year: int, acronym: str, person: Person) -> List:
+        configuration = learning_unit_sdk.build_configuration()
+        with osis_learning_unit_sdk.ApiClient(configuration) as api_client:
+            api_instance = learning_units_api.LearningUnitsApi(api_client)
+            try:
+                learning_unit_title = api_instance.learningunitstitle_read(
+                    year=year,
+                    acronym=acronym,
+                    **utils.build_mandatory_auth_headers(person),
+                )['title']
+            except (osis_learning_unit_sdk.ApiException, urllib3.exceptions.HTTPError,) as e:
+                # Run in degraded mode in order to prevent crash all app
+                logger.error(e)
+                learning_unit_title = ''
+        return learning_unit_title
+
+    @staticmethod
+    def get_effective_classes(year: int, acronym: str, person: Person) -> List[EffectiveClass]:
+        configuration = learning_unit_sdk.build_configuration()
+        with osis_learning_unit_sdk.ApiClient(configuration) as api_client:
+            api_instance = learning_units_api.LearningUnitsApi(api_client)
+            try:
+                effective_classes = api_instance.get_learning_unit_classes(
+                    year=year,
+                    acronym=acronym,
+                    **utils.build_mandatory_auth_headers(person),
+                )
+            except (osis_learning_unit_sdk.ApiException, urllib3.exceptions.HTTPError,) as e:
+                # Run in degraded mode in order to prevent crash all app
+                logger.error(e)
+                effective_classes = []
+        return effective_classes
